Moving datastore GAX/gRPC helpers out of _http and into dedicated module.
